Including new methods for producers API.
